Added api documentation on /api/documentation and /api/documentation.pdf
